Refactored out code that will be used on robot.
