+++ app/src/main/java/org/firstinspires/ftc/teamcode/PolePosition.java
@@ -1,4 +1,4 @@
-package opencv_test;
+package org.firstinspires.ftc.teamcode;
 
 public class PolePosition {
     private int width;
